% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{censo_borrar_base}
\alias{censo_borrar_base}
\title{Elimina la Base de Datos del Censo de tu Computador}
\usage{
censo_borrar_base()
}
\description{
Elimina el directorio \code{censo2017} en el directorio de datos de usuario.
}
\examples{
\dontrun{ censo_borrar_base() }
}
