% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{censo_datos}
\alias{censo_datos}
\title{Tablas Completas de la Base de Datos del Censo}
\usage{
censo_datos(tabla)
}
\arguments{
\item{tabla}{Una cadena de texto indicando la tabla a extraer}
}
\value{
Un tibble
}
\description{
Devuelve una tabla completa de la base de datos. Para entregar datos
filtrados previamente se debe usar \code{\link[=censo_bbdd]{censo_bbdd()}}. Esta función puede
ser especialmente útil para obtener los mapas y usarlos directamente con
tm o ggplot2, sin necesidad de transformar las columnas de geometrías.
}
\examples{
if (censo_estado()) {
  censo_datos("comunas")
}
}
