% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{censo_descargar}
\alias{censo_descargar}
\title{Descarga la Base de Datos del Censo a tu Computador}
\usage{
censo_descargar(ver = NULL)
}
\arguments{
\item{ver}{La version a descargar. Por defecto es la ultima version
disponible en GitHub. Se pueden ver todas las versiones en
\url{https://github.com/pachamaltese/censo2017/releases}.}
}
\description{
Este comando descarga la base de datos completa como un unico archivo zip que
se descomprime para crear la base de datos local. Si no quieres descargar la
base de datos en tu home, ejecuta usethis::edit_r_environ() para crear la
variable de entorno CENSO2017_DIR con la ruta.
}
\examples{
\dontrun{ censo_descargar() }
}
