% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{censo_desconectar}
\alias{censo_desconectar}
\title{Desconecta la Base de Datos del Censo}
\usage{
censo_desconectar()
}
\description{
Una funcion auxiliar para desconectarse de la base de datos.
}
\examples{
censo_desconectar()
}
