% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{censo_eliminar}
\alias{censo_eliminar}
\title{Elimina la Base de Datos del Censo de tu Computador}
\usage{
censo_eliminar()
}
\description{
Elimina el directorio \code{censo2017} y todos sus contenidos, incluyendo versiones
de la base de datos del Censo creadas con cualquier version de 'DuckDB'.
}
\examples{
\dontrun{ censo_eliminar() }
}
