% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{censo_estado}
\alias{censo_estado}
\title{Obtiene el Estado de la Base de Datos Local del Censo}
\usage{
censo_estado(msg = TRUE)
}
\arguments{
\item{msg}{Mostrar o no mensajes de estado. Por defecto es TRUE.}
}
\value{
TRUE si la base de datos existe y contiene las tablas esperadas,
FALSE  en caso contrario (invisible).
}
\description{
Entrega el estado de la base de datos local. Muestra un mensaje informativo
respecto de como obtener la base si esta no se encuentra o esta daniada.
}
\examples{
censo_estado()
}
